Adding burn contract. (#585)
## 1. Overview

In this PR we are adding Burn contract that will burn the token sent to
it.

## 2. Implementation details

The Burn Coins Contract is a simple smart contract designed to burn the
coins sent to it. Here are some highlights of the implementation:

- The contract consists of an `instantiate` function to initialize the
contract.
- It includes an `execute` function that has `burn` function to handle
the execution of coin burning.
- The contract allows to query the total amount of coins burnt using the
`query` function.

Design Rationale:
- The contract follows a straightforward design to achieve its primary
objective of burning coins.
- It ensures that burning coins is irreversible to maintain the
integrity of the tokenomics.

## 3. How to test/use

To test or use the Burn Coins Contract, follow these steps:
- Build the contract locally
- Change to demos directory
- Run `sh quasar_localnet.sh`
- Run `burn.sh`

## 4. Checklist

- [x] Does the Readme need to be updated?

## 5. Limitations (optional)

<!-- Describe any limitation of the capabilities listed in the Overview
section. -->

## 6. Future Work (optional)

Potential future enhancements for the Burn Coins Contract include:
- Implementing additional functionality for managing burnt coins, such
as tracking burnt amounts per user or token.
- Adding support for configurable burning parameters, such as burn fees
or burn rates.

---------

Co-authored-by: Laurens <32776056+LaurensKubat@users.noreply.github.com>