Feat/events refund (#648)
This PR aims to revert the changes to the `refund0` and `refund1`
attributes from a previous PR that broke our Subgraph indexer. I patched
and reingested the Subgraph, but since this seems redundant and
unnecessary, I would like to revert to the previous version.
`refund0_denom` and `refund1_denom` are useless, as 0 and 1 already
provide clients with sufficient information, and it is inconsistent with
how we still treat `amount0` and `amount1`.

Minor changes include moving helpers to the helpers mod.